Don't BuildRequire pyflakes
Linting code is extremely useful in upstream, but not so much in downstream.

When we update Python to 3.N+1, linters are often broken for a while and pycurl
is part of the initial Python bootstrap sequence.
Not linting the code in %check allows us to rebuild pycurl without functional
pyflakes. See https://bugzilla.redhat.com/show_bug.cgi?id=1831248